third_party/WebKit/LayoutTests/fast/dom/StyleSheet/detached-parent-rule-without-wrapper-expected.txt

Test that detached CSSOM subtrees keeps a reference to CSS stylesheets lacking wrappers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getStyleSheetRules().length is 1
Removing <style>...
PASS rule.parentRule is not null
PASS successfullyParsed is true

TEST COMPLETE
